+++ Kernel.py
@@ -1,9 +1,10 @@
 import os
+import re
 import time
 from threading import Thread, Lock
 
-from LaserOperation import *
-from svgelements import *
+from LaserCommandConstants import *
+from svgelements import Color
 from zinglplotter import ZinglPlotter
 
 STATE_UNKNOWN = -1
@@ -556,2143 +557,6 @@ class Spooler(Modifier):
         self.context.signal('spooler;queue', len(self._queue))
 
 
-class Elemental(Modifier):
-    """
-    The elemental module is governs all the interactions with the various elements,
-    operations, and filenodes. Handling structure change and selection, emphasis, and
-    highlighting changes. The goal of this module is to make sure that the life cycle
-    of the elements is strictly enforced. For example, every element that is removed
-    must have had the .cache deleted. And anything selecting an element must propagate
-    that information out to inform other interested modules.
-    """
-
-    def __init__(self, context, name=None, channel=None, *args, **kwargs):
-        Modifier.__init__(self, context, name, channel)
-        self._plan = list()
-        self._operations = list()
-        self._elements = list()
-        self._filenodes = {}
-        self.note = None
-        self._bounds = None
-
-    def attach(self, channel=None):
-        context = self.context
-        context.elements = self
-        context.classify = self.classify
-        context.save = self.save
-        context.save_types = self.save_types
-        context.load = self.load
-        context.load_types = self.load_types
-        self.add_elem(Rect(0,0,1000,1000, stroke='black'))
-
-        context = self.context
-        kernel = self.context._kernel
-        elements = self
-        _ = kernel.translation
-
-        def grid(command, *args):
-            try:
-                cols = int(args[0])
-                rows = int(args[1])
-                x_distance = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-                y_distance = Length(args[3]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            except (ValueError, IndexError):
-                yield _("Syntax Error: grid <columns> <rows> <x_distance> <y_distance>")
-                return
-            items = list(elements.elems(emphasized=True))
-            y_pos = elements._bounds[1]
-            for j in range(rows):
-                x_pos = elements._bounds[0]
-                for k in range(cols):
-                    if j != 0 or k != 0:
-                        add_elem = list(map(copy, items))
-                        for e in add_elem:
-                            e *= 'translate(%f, %f)' % (x_pos, y_pos)
-                        elements.add_elems(add_elem)
-                    x_pos += x_distance
-                y_pos += y_distance
-
-        kernel.register('command/grid', grid)
-
-        def element(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Graphical Elements:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    if element.emphasized:
-                        yield '%d: * %s' % (i, name)
-                    else:
-                        yield '%d: %s' % (i, name)
-                    i += 1
-                yield _('----------')
-            else:
-                for value in args:
-                    try:
-                        value = int(value)
-                    except ValueError:
-                        if value == "*":
-                            yield _('Selecting all elements.')
-                            elements.set_selected(list(elements.elems()))
-                            continue
-                        elif value == "~":
-                            yield _('Invert selection.')
-                            elements.set_selected(list(elements.elems(emphasized=False)))
-                            continue
-                        elif value == "!":
-                            yield _('Select none')
-                            elements.set_selected(None)
-                            continue
-                        elif value == "delete":
-                            yield _('deleting.')
-                            elements.remove_elements(list(elements.elems(emphasized=True)))
-                            self.context.signal('refresh_scene', 0)
-                            continue
-                        elif value == "copy":
-                            add_elem = list(map(copy, elements.elems(emphasized=True)))
-                            elements.add_elems(add_elem)
-                            for e in add_elem:
-                                e.select()
-                                e.emphasize()
-                            continue
-                        elif value == "merge":
-                            superelement = Path()
-                            for e in elements.elems(emphasized=True):
-                                if superelement.stroke is None:
-                                    superelement.stroke = e.stroke
-                                if superelement.fill is None:
-                                    superelement.fill = e.fill
-                                superelement += abs(e)
-                            elements.remove_elements(list(elements.elems(emphasized=True)))
-                            elements.add_elem(superelement)
-                            superelement.emphasize()
-                            continue
-                        elif value == "subpath":
-                            for e in elements.elems(emphasized=True):
-                                p = abs(e)
-                                add = []
-                                for subpath in p.as_subpaths():
-                                    subelement = Path(subpath)
-                                    add.append(subelement)
-                                elements.add_elems(add)
-                            continue
-                        yield _('Value Error: %s is not an integer') % value
-                        continue
-                    try:
-                        element = elements.get_elem(value)
-                        name = str(element)
-                        if len(name) > 50:
-                            name = name[:50] + '...'
-                        if element.selected:
-                            element.unselect()
-                            element.unemphasize()
-                            yield _('Deselecting item %d called %s') % (value, name)
-                        else:
-                            element.select()
-                            element.emphasize()
-                            yield _('Selecting item %d called %s') % (value, name)
-                    except IndexError:
-                        yield _('index %d out of range') % value
-            return
-
-        kernel.register('command/element', element)
-
-        def path(command, *args):
-            path_d = ' '.join(args)
-            element = Path(path_d)
-            self.add_element(element)
-            return
-
-        kernel.register('command/path', path)
-
-        def circle(command, *args):
-            if len(args) == 3:
-                x_pos = Length(args[0]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-                y_pos = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-                r_pos = Length(args[2]).value(ppi=1000.0,
-                                              relative_length=min(self.context.bed_height,
-                                                                  self.context.bed_width) * 39.3701)
-            elif len(args) == 1:
-                x_pos = 0
-                y_pos = 0
-                r_pos = Length(args[0]).value(ppi=1000.0,
-                                              relative_length=min(self.context.bed_height,
-                                                                  self.context.bed_width) * 39.3701)
-            else:
-                yield _('Circle <x> <y> <r> or circle <r>')
-                return
-            element = Circle(cx=x_pos, cy=y_pos, r=r_pos)
-            element = Path(element)
-            self.add_element(element)
-
-        kernel.register('command/circle', circle)
-
-        def ellipse(command, *args):
-            if len(args) < 4:
-                yield _('Too few arguments (needs center_x, center_y, radius_x, radius_y)')
-                return
-            x_pos = Length(args[0]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            y_pos = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            rx_pos = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            ry_pos = Length(args[3]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            element = Ellipse(cx=x_pos, cy=y_pos, rx=rx_pos, ry=ry_pos)
-            element = Path(element)
-            self.add_element(element)
-            return
-
-        kernel.register('command/ellipse', ellipse)
-
-        def rect(command, *args):
-            if len(args) < 4:
-                yield _("Too few arguments (needs x, y, width, height)")
-                return
-            x_pos = Length(args[0]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            y_pos = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            width = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            height = Length(args[3]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            element = Rect(x=x_pos, y=y_pos, width=width, height=height)
-            element = Path(element)
-            self.add_element(element)
-            return
-
-        kernel.register('command/rect', rect)
-
-        def text(command, *args):
-            text = ' '.join(args)
-            element = SVGText(text)
-            self.add_element(element)
-
-        kernel.register('command/text', text)
-
-        def polygon(command, *args):
-            element = Polygon(list(map(float, args)))
-            element = Path(element)
-            self.add_element(element)
-            return
-
-        kernel.register('command/polygon', polygon)
-
-        def polyline(command, *args):
-            element = Polygon(list(map(float, args)))
-            element = Path(element)
-            self.add_element(element)
-            return
-
-        kernel.register('command/polyline', polyline)
-
-        def stroke(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Stroke Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    if element.stroke is None or element.stroke == 'none':
-                        yield _('%d: stroke = none - %s') % (i, name)
-                    else:
-                        yield _('%d: stroke = %s - %s') % (i, element.stroke.hex, name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _("No selected elements.")
-                return
-            if args[0] == 'none':
-                for element in elements.elems(emphasized=True):
-                    element.stroke = None
-                    element.altered()
-            else:
-                for element in elements.elems(emphasized=True):
-                    element.stroke = Color(args[0])
-                    element.altered()
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/stroke', stroke)
-
-        def fill(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Fill Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    if element.fill is None or element.fill == 'none':
-                        yield _('%d: fill = none - %s') % (i, name)
-                    else:
-                        yield _('%d: fill = %s - %s') % (i, element.fill.hex, name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            if args[0] == 'none':
-                for element in elements.elems(emphasized=True):
-                    element.fill = None
-                    element.altered()
-            else:
-                for element in elements.elems(emphasized=True):
-                    element.fill = Color(args[0])
-                    element.altered()
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/fill', fill)
-
-        def rotate(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Rotate Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield _('%d: rotate(%fturn) - %s') % \
-                          (i, element.rotation.as_turns, name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            bounds = elements.bounds()
-            if len(args) >= 1:
-                rot = Angle.parse(args[0]).as_degrees
-            else:
-                rot = 0
-            if len(args) >= 2:
-                center_x = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_x = (bounds[2] + bounds[0]) / 2.0
-            if len(args) >= 3:
-                center_y = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_y = (bounds[3] + bounds[1]) / 2.0
-            matrix = Matrix('rotate(%f,%f,%f)' % (rot, center_x, center_y))
-            try:
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            continue
-                    except AttributeError:
-                        pass
-
-                    element *= matrix
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/rotate', rotate)
-
-        def scale(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Scale Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield '%d: scale(%f, %f) - %s' % \
-                          (i, element.transform.value_scale_x(), element.transform.value_scale_x(), name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            bounds = elements.bounds()
-
-            if len(args) >= 1:
-                sx = Length(args[0]).value(relative_length=1.0)
-            else:
-                sx = 1
-            if len(args) >= 2:
-                sy = Length(args[1]).value(relative_length=1.0)
-            else:
-                sy = sx
-            if len(args) >= 3:
-                center_x = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_x = (bounds[2] + bounds[0]) / 2.0
-            if len(args) >= 4:
-                center_y = Length(args[3]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_y = (bounds[3] + bounds[1]) / 2.0
-            if sx == 0 or sy == 0:
-                yield _('Scaling by Zero Error')
-                return
-            matrix = Matrix('scale(%f,%f,%f,%f)' % (sx, sy, center_x, center_y))
-            try:
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            continue
-                    except AttributeError:
-                        pass
-
-                    element *= matrix
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/scale', scale)
-
-        def translate(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Translate Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield _('%d: translate(%f, %f) - %s') % \
-                          (i, element.transform.value_trans_x(), element.transform.value_trans_y(), name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            if len(args) >= 1:
-                tx = Length(args[0]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                tx = 0
-            if len(args) >= 2:
-                ty = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                ty = 0
-            matrix = Matrix('translate(%f,%f)' % (tx, ty))
-            try:
-                for element in elements.elems(emphasized=True):
-                    element *= matrix
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/translate', translate)
-
-        def rotate_to(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Rotate Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield _('%d: rotate(%fturn) - %s') % \
-                          (i, element.rotation.as_turns, name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            bounds = elements.bounds()
-            try:
-                end_angle = Angle.parse(args[0])
-            except ValueError:
-                yield _('Invalid Value.')
-                return
-            if len(args) >= 2:
-                center_x = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_x = (bounds[2] + bounds[0]) / 2.0
-            if len(args) >= 3:
-                center_y = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_y = (bounds[3] + bounds[1]) / 2.0
-
-            try:
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            continue
-                    except AttributeError:
-                        pass
-
-                    start_angle = element.rotation
-                    amount = end_angle - start_angle
-                    matrix = Matrix('rotate(%f,%f,%f)' % (Angle(amount).as_degrees, center_x, center_y))
-                    element *= matrix
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/rotate_to', rotate_to)
-
-        def scale_to(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Scale Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield _('%d: scale(%f, %f) - %s') % \
-                          (i, element.transform.value_scale_x(), element.transform.value_scale_y(), name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            bounds = elements.bounds()
-            if len(args) >= 1:
-                sx = Length(args[0]).value(relative_length=1.0)
-            else:
-                sx = 1
-            if len(args) >= 2:
-                sy = Length(args[1]).value(relative_length=1.0)
-            else:
-                sy = sx
-            if len(args) >= 3:
-                center_x = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_x = (bounds[2] + bounds[0]) / 2.0
-            if len(args) >= 4:
-                center_y = Length(args[3]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                center_y = (bounds[3] + bounds[1]) / 2.0
-            try:
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            continue
-                    except AttributeError:
-                        pass
-
-                    osx = element.transform.value_scale_x()
-                    osy = element.transform.value_scale_y()
-                    if sx == 0 or sy == 0:
-                        yield _('Scaling by Zero Error')
-                        return
-                    nsx = sx / osx
-                    nsy = sy / osy
-                    matrix = Matrix('scale(%f,%f,%f,%f)' % (nsx, nsy, center_x, center_y))
-                    element *= matrix
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/scale_to', scale_to)
-
-        def translate_to(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Translate Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield _('%d: translate(%f, %f) - %s') % \
-                          (i, element.transform.value_trans_x(), element.transform.value_trans_y(), name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-
-            if len(args) >= 1:
-                tx = Length(args[0]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-            else:
-                tx = 0
-            if len(args) >= 2:
-                ty = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            else:
-                ty = 0
-            try:
-                for element in elements.elems(emphasized=True):
-                    otx = element.transform.value_trans_x()
-                    oty = element.transform.value_trans_y()
-                    ntx = tx - otx
-                    nty = ty - oty
-                    matrix = Matrix('translate(%f,%f)' % (ntx, nty))
-                    element *= matrix
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/translate_to', translate_to)
-
-        def resize(command, *args):
-            if len(args) < 4:
-                yield _('Too few arguments (needs x, y, width, height)')
-                return
-            try:
-                x_pos = Length(args[0]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701)
-                y_pos = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-                w_dim = Length(args[2]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-                h_dim = Length(args[3]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-                x, y, x1, y1 = elements.bounds()
-                w, h = x1 - x, y1 - y
-                sx = w_dim / w
-                sy = h_dim / h
-                matrix = Matrix('translate(%f,%f) scale(%f,%f) translate(%f,%f)' % (x_pos, y_pos, sx, sy, -x, -y))
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            continue
-                    except AttributeError:
-                        pass
-                    element *= matrix
-                    element.modified()
-                context.signal('refresh_scene')
-            except (ValueError, ZeroDivisionError):
-                return
-
-        kernel.register('command/resize', resize)
-
-        def matrix(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Matrix Values:')
-                i = 0
-                for element in elements.elems():
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield '%d: %s - %s' % \
-                          (i, str(element.transform), name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            if len(args) != 6:
-                yield _('Requires six matrix parameters')
-                return
-            try:
-                matrix = Matrix(float(args[0]), float(args[1]), float(args[2]), float(args[3]),
-                                Length(args[4]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701),
-                                Length(args[5]).value(ppi=1000.0, relative_length=self.context.bed_width * 39.3701))
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            continue
-                    except AttributeError:
-                        pass
-
-                    element.transform = Matrix(matrix)
-                    element.modified()
-            except ValueError:
-                yield _('Invalid value')
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/matrix', matrix)
-
-        def reset(command, *args):
-            for element in elements.elems(emphasized=True):
-                try:
-                    if element.lock:
-                        continue
-                except AttributeError:
-                    pass
-
-                name = str(element)
-                if len(name) > 50:
-                    name = name[:50] + '...'
-                yield _('reset - %s') % name
-                element.transform.reset()
-                element.modified()
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/reset', reset)
-
-        def reify(command, *args):
-            for element in elements.elems(emphasized=True):
-                try:
-                    if element.lock:
-                        continue
-                except AttributeError:
-                    pass
-
-                name = str(element)
-                if len(name) > 50:
-                    name = name[:50] + '...'
-                yield _('reified - %s') % name
-                element.reify()
-                element.altered()
-            context.signal('refresh_scene')
-            return
-
-        kernel.register('command/reify', reify)
-
-        # REQUIRES CUTPLANNER
-        def optimize(command, *args):
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            elif len(args) == 0:
-                yield _('Optimizations: cut_inner, travel, cut_travel')
-                return
-            elif args[0] == 'cut_inner':
-                for element in elements.elems(emphasized=True):
-                    e = CutPlanner.optimize_cut_inside(element)
-                    element.clear()
-                    element += e
-                    element.altered()
-            elif args[0] == 'travel':
-                yield _('Travel Optimizing: %f') % CutPlanner.length_travel(elements.elems(emphasized=True))
-                for element in elements.elems(emphasized=True):
-                    e = CutPlanner.optimize_travel(element)
-                    element.clear()
-                    element += e
-                    element.altered()
-                yield _('Optimized: %f') % CutPlanner.length_travel(elements.elems(emphasized=True))
-            elif args[0] == 'cut_travel':
-                yield _('Cut Travel Initial: %f') % CutPlanner.length_travel(elements.elems(emphasized=True))
-                for element in elements.elems(emphasized=True):
-                    e = CutPlanner.optimize_general(element)
-                    element.clear()
-                    element += e
-                    element.altered()
-                yield _('Cut Travel Optimized: %f') % CutPlanner.length_travel(elements.elems(emphasized=True))
-            else:
-                yield _('Optimization not found.')
-                return
-
-        kernel.register('command/optimize', optimize)
-
-        # REQUIRES CUTPLANNER
-        def embroider(command, *args):
-            yield _('Embroidery Filling')
-            if len(args) >= 1:
-                angle = Angle.parse(args[0])
-            else:
-                angle = None
-            if len(args) >= 2:
-                distance = Length(args[1]).value(ppi=1000.0, relative_length=self.context.bed_height * 39.3701)
-            else:
-                distance = 16
-            for element in elements.elems(emphasized=True):
-                if not isinstance(element, Path):
-                    continue
-                if angle is not None:
-                    element *= Matrix.rotate(angle)
-                e = CutPlanner.eulerian_fill([abs(element)], distance=distance)
-                element.transform.reset()
-                element.clear()
-                element += e
-                if angle is not None:
-                    element *= Matrix.rotate(-angle)
-                element.altered()
-
-        kernel.register('command/embroider', embroider)
-
-        def operation(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Operations:')
-                i = 0
-
-                for operation in elements.ops():
-                    selected = operation.selected
-                    name = str(operation)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    if selected:
-                        yield '%d: * %s' % (i, name)
-                    else:
-                        yield '%d: %s' % (i, name)
-                    i += 1
-                yield _('----------')
-            else:
-                for value in args:
-                    try:
-                        value = int(value)
-                    except ValueError:
-                        if value == "*":
-                            yield _('Selecting all operations.')
-                            elements.set_selected(list(elements.ops()))
-                            continue
-                        elif value == "~":
-                            yield _('Invert selection.')
-                            elements.set_selected(list(elements.ops(emphasized=False)))
-                            continue
-                        elif value == "!":
-                            yield _('Select none')
-                            elements.set_selected(None)
-                            continue
-                        elif value == "delete":
-                            yield _('Deleting.')
-                            elements.remove_operations(list(elements.ops(emphasized=True)))
-                            continue
-                        elif value == "copy":
-                            add_elem = list(map(copy, elements.ops(emphasized=True)))
-                            elements.add_ops(add_elem)
-                            for e in add_elem:
-                                e.select()
-                                e.emphasize()
-                            continue
-                        yield _('Value Error: %s is not an integer') % value
-                        continue
-                    try:
-                        operation = elements.get_op(value)
-                        name = str(operation)
-                        if len(name) > 50:
-                            name = name[:50] + '...'
-                        if operation.emphasized:
-                            operation.unemphasize()
-                            operation.unselect()
-                            yield _('Deselecting operation %d called %s') % (value, name)
-                        else:
-                            operation.emphasize()
-                            operation.select()
-                            yield _('Selecting operation %d called %s') % (value, name)
-                    except IndexError:
-                        yield _('index %d out of range') % value
-            return
-
-        kernel.register('command/operation', operation)
-
-        def classify(command, *args):
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            elements.classify(list(elements.elems(emphasized=True)))
-            return
-
-        kernel.register('command/classify', classify)
-
-        def declassify(command, *args):
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            elements.remove_elements_from_operations(list(elements.elems(emphasized=True)))
-            return
-
-        kernel.register('command/declassify', declassify)
-
-        def note(command, *args):
-            if len(args) == 0:
-                if elements.note is None:
-                    yield _('No Note.')
-                else:
-                    yield str(elements.note)
-            else:
-                elements.note = ' '.join(args)
-                yield _('Note Set.')
-
-        kernel.register('command/note', note)
-
-        def cut(command, *args):
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            op = LaserOperation()
-            op.operation = "Cut"
-            op.extend(elements.elems(emphasized=True))
-            elements.add_op(op)
-            return
-
-        kernel.register('command/cut', cut)
-
-        def engrave(command, *args):
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            op = LaserOperation()
-            op.operation = "Engrave"
-            op.extend(elements.elems(emphasized=True))
-            elements.add_op(op)
-            return
-
-        kernel.register('command/engrave', engrave)
-
-        def raster(command, *args):
-            if not elements.has_emphasis():
-                yield _('No selected elements.')
-                return
-            op = LaserOperation()
-            op.operation = "Raster"
-            op.extend(elements.elems(emphasized=True))
-            elements.add_op(op)
-            return
-
-        kernel.register('command/raster', raster)
-
-        def step(command, *args):
-            if len(args) == 0:
-                found = False
-                for op in elements.ops(emphasized=True):
-                    if op.operation in ("Raster", "Image"):
-                        step = op.raster_step
-                        yield _('Step for %s is currently: %d') % (str(op), step)
-                        found = True
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        try:
-                            step = element.values['raster_step']
-                        except KeyError:
-                            step = 1
-                        yield _('Image step for %s is currently: %s') % (str(element), step)
-                        found = True
-                if not found:
-                    yield _('No raster operations selected.')
-                return
-            try:
-                step = int(args[0])
-            except ValueError:
-                yield _('Not integer value for raster step.')
-                return
-            for op in elements.ops(emphasized=True):
-                if op.operation in ("Raster", "Image"):
-                    op.raster_step = step
-                    self.context.signal('element_property_update', op)
-            for element in elements.elems(emphasized=True):
-                element.values['raster_step'] = str(step)
-                m = element.transform
-                tx = m.e
-                ty = m.f
-                element.transform = Matrix.scale(float(step), float(step))
-                element.transform.post_translate(tx, ty)
-                element.modified()
-                self.context.signal('element_property_update', element)
-                self.context.signal('refresh_scene')
-            return
-
-        kernel.register('command/step', step)
-
-        # REQUIRES OPERATION PREPROCESSOR.
-        def image(command, *args):
-            if len(args) == 0:
-                yield _('----------')
-                yield _('Images:')
-                i = 0
-                for element in elements.elems():
-                    if not isinstance(element, SVGImage):
-                        continue
-                    name = str(element)
-                    if len(name) > 50:
-                        name = name[:50] + '...'
-                    yield '%d: (%d, %d) %s, %s' % (i,
-                                                   element.image_width,
-                                                   element.image_height,
-                                                   element.image.mode,
-                                                   name)
-                    i += 1
-                yield _('----------')
-                return
-            if not elements.has_emphasis():
-                yield _('No selected images.')
-                return
-            elif args[0] == 'wizard':
-                if len(args) == 1:
-                    try:
-                        for script_name in context.match('raster_script', True):
-                            yield _('Raster Script: %s') % script_name
-                    except KeyError:
-                        yield _('No Raster Scripts Found.')
-                    return
-                try:
-                    script = context.registered['raster_script/%s' % args[1]]
-                except KeyError:
-                    yield _('Raster Script %s is not registered.') % args[1]
-                    return
-                from RasterScripts import RasterScripts
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        element.image, element.transform, step = RasterScripts.wizard_image(element, script)
-                        if step is not None:
-                            element.values['raster_step'] = step
-                        element.image_width, element.image_height = element.image.size
-                        element.lock = True
-                        element.altered()
-                return
-            elif args[0] == 'unlock':
-                yield _('Unlocking Elements...')
-                for element in elements.elems(emphasized=True):
-                    try:
-                        if element.lock:
-                            yield "Unlocked: %s" % str(element)
-                            element.lock = False
-                        else:
-                            yield _('Element was not locked: %s') % str(element)
-                    except AttributeError:
-                        yield _('Element was not locked: %s') % str(element)
-                return
-            elif args[0] == 'threshold':
-                try:
-                    threshold_min = float(args[1])
-                    threshold_max = float(args[2])
-                except (ValueError, IndexError):
-                    yield _('Threshold values improper.')
-                    return
-                divide = (threshold_max - threshold_min) / 255.0
-                for element in elements.elems(emphasized=True):
-                    if not isinstance(element, SVGImage):
-                        continue
-                    image_element = copy(element)
-                    image_element.image = image_element.image.copy()
-                    if OperationPreprocessor.needs_actualization(image_element):
-                        OperationPreprocessor.make_actual(image_element)
-                    img = image_element.image
-                    new_data = img.load()
-                    width, height = img.size
-                    for y in range(height):
-                        for x in range(width):
-                            pixel = new_data[x, y]
-                            if pixel[3] == 0:
-                                new_data[x, y] = (255, 255, 255, 255)
-                                continue
-                            gray = (pixel[0] + pixel[1] + pixel[2]) / 3.0
-                            if threshold_min >= gray:
-                                new_data[x, y] = (0, 0, 0, 255)
-                            elif threshold_max < gray:
-                                new_data[x, y] = (255, 255, 255, 255)
-                            else:  # threshold_min <= grey < threshold_max
-                                v = gray - threshold_min
-                                v *= divide
-                                v = int(round(v))
-                                new_data[x, y] = (v, v, v, 255)
-                    elements.add_elem(image_element)
-            elif args[0] == 'resample':
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        OperationPreprocessor.make_actual(element)
-                        element.altered()
-                return
-            elif args[0] == 'dither':
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        if img.mode == 'RGBA':
-                            pixel_data = img.load()
-                            width, height = img.size
-                            for y in range(height):
-                                for x in range(width):
-                                    if pixel_data[x, y][3] == 0:
-                                        pixel_data[x, y] = (255, 255, 255, 255)
-                        element.image = img.convert("1")
-                        element.altered()
-            elif args[0] == 'remove':
-                if len(args) == 1:
-                    yield _('Must specify a color, and optionally a distance.')
-                    return
-                distance = 50.0
-                color = "White"
-                if len(args) >= 2:
-                    color = args[1]
-                try:
-                    color = Color(color)
-                except ValueError:
-                    yield _('Color Invalid.')
-                    return
-                if len(args) >= 3:
-                    try:
-                        distance = float(args[2])
-                    except ValueError:
-                        yield _('Color distance is invalid.')
-                        return
-                distance_sq = distance * distance
-
-                def dist(pixel):
-                    r = color.red - pixel[0]
-                    g = color.green - pixel[1]
-                    b = color.blue - pixel[2]
-                    return r * r + g * g + b * b <= distance_sq
-
-                for element in elements.elems(emphasized=True):
-                    if not isinstance(element, SVGImage):
-                        continue
-                    img = element.image
-                    if img.mode != "RGBA":
-                        img = img.convert('RGBA')
-                    new_data = img.load()
-                    width, height = img.size
-                    for y in range(height):
-                        for x in range(width):
-                            pixel = new_data[x, y]
-                            if dist(pixel):
-                                new_data[x, y] = (255, 255, 255, 0)
-                                continue
-                    element.image = img
-                    element.altered()
-            elif args[0] == 'add':
-                if len(args) == 1:
-                    yield _('Must specify a color, to add.')
-                    return
-                color = "White"
-                if len(args) >= 2:
-                    color = args[1]
-                try:
-                    color = Color(color)
-                except ValueError:
-                    yield _('Color Invalid.')
-                    return
-                pix = (color.red, color.green, color.blue, color.alpha)
-                for element in elements.elems(emphasized=True):
-                    if not isinstance(element, SVGImage):
-                        continue
-                    img = element.image
-                    if img.mode != "RGBA":
-                        img = img.convert('RGBA')
-                    new_data = img.load()
-                    width, height = img.size
-                    for y in range(height):
-                        for x in range(width):
-                            pixel = new_data[x, y]
-                            if pixel[3] == 0:
-                                new_data[x, y] = pix
-                                continue
-                    element.image = img
-                    element.altered()
-            elif args[0] == 'crop':
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        try:
-                            left = int(args[1])
-                            upper = int(args[2])
-                            right = int(args[3])
-                            lower = int(args[4])
-                            element.image = img.crop((left, upper, right, lower))
-                            element.image_width = right - left
-                            element.image_height = lower - upper
-                            element.altered()
-                        except (KeyError, ValueError):
-                            yield _('image crop <left> <upper> <right> <lower>')
-                return
-            elif args[0] == 'contrast':
-                for element in elements.elems(emphasized=True):
-                    from PIL import ImageEnhance
-                    if isinstance(element, SVGImage):
-                        try:
-
-                            factor = float(args[1])
-                            img = element.image
-                            enhancer = ImageEnhance.Contrast(img)
-                            element.image = enhancer.enhance(factor)
-                            element.altered()
-                            yield _('Image Contrast Factor: %f') % factor
-                        except (IndexError, ValueError):
-                            yield _('image contrast <factor>')
-                return
-            elif args[0] == 'brightness':
-                for element in elements.elems(emphasized=True):
-                    from PIL import ImageEnhance
-                    if isinstance(element, SVGImage):
-                        try:
-                            factor = float(args[1])
-                            img = element.image
-                            enhancer = ImageEnhance.Brightness(img)
-                            element.image = enhancer.enhance(factor)
-                            element.altered()
-                            yield _('Image Brightness Factor: %f') % factor
-                        except (IndexError, ValueError):
-                            yield _('image brightness <factor>')
-                return
-            elif args[0] == 'color':
-                for element in elements.elems(emphasized=True):
-                    from PIL import ImageEnhance
-                    if isinstance(element, SVGImage):
-                        try:
-                            factor = float(args[1])
-                            img = element.image
-                            enhancer = ImageEnhance.Color(img)
-                            element.image = enhancer.enhance(factor)
-                            element.altered()
-                            yield _('Image Color Factor: %f') % factor
-                        except (IndexError, ValueError):
-                            yield _('image color <factor>')
-                return
-            elif args[0] == 'sharpness':
-                for element in elements.elems(emphasized=True):
-                    from PIL import ImageEnhance
-                    if isinstance(element, SVGImage):
-                        try:
-                            factor = float(args[1])
-                            img = element.image
-                            enhancer = ImageEnhance.Sharpness(img)
-                            element.image = enhancer.enhance(factor)
-                            element.altered()
-                            yield _('Image Sharpness Factor: %f') % factor
-                        except (IndexError, ValueError):
-                            yield _('image sharpness <factor>')
-                return
-            elif args[0] == 'blur':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.BLUR)
-                        element.altered()
-                        yield _('Image Blurred.')
-                return
-            elif args[0] == 'sharpen':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.SHARPEN)
-                        element.altered()
-                        yield _('Image Sharpened.')
-                return
-            elif args[0] == 'edge_enhance':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.EDGE_ENHANCE)
-                        element.altered()
-                        yield _('Image Edges Enhanced.')
-                return
-            elif args[0] == 'find_edges':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.FIND_EDGES)
-                        element.altered()
-                        yield _('Image Edges Found.')
-                return
-            elif args[0] == 'emboss':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.EMBOSS)
-                        element.altered()
-                        yield _('Image Embossed.')
-                return
-            elif args[0] == 'smooth':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.SMOOTH)
-                        element.altered()
-                        yield _('Image Smoothed.')
-                return
-            elif args[0] == 'contour':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.CONTOUR)
-                        element.altered()
-                        yield _('Image Contoured.')
-                return
-            elif args[0] == 'detail':
-                from PIL import ImageFilter
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.filter(filter=ImageFilter.DETAIL)
-                        element.altered()
-                        yield _('Image Detailed.')
-                return
-            elif args[0] == 'quantize':
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        try:
-                            colors = int(args[1])
-                            img = element.image
-                            element.image = img.quantize(colors=colors)
-                            element.altered()
-                            yield _('Image Quantized to %d colors.') % colors
-                        except (IndexError, ValueError):
-                            yield _('image quantize <colors>')
-                return
-            elif args[0] == 'solarize':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        try:
-                            threshold = int(args[1])
-                            img = element.image
-                            element.image = ImageOps.solarize(img, threshold=threshold)
-                            element.altered()
-                            yield _('Image Solarized at %d gray.') % threshold
-                        except (IndexError, ValueError):
-                            yield _('image solarize <threshold>')
-                return
-            elif args[0] == 'invert':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        original_mode = img.mode
-                        if img.mode == 'P' or img.mode == 'RGBA' or img.mode == '1':
-                            img = img.convert('RGB')
-                        try:
-                            element.image = ImageOps.invert(img)
-                            if original_mode == '1':
-                                element.image = element.image.convert('1')
-                            element.altered()
-                            yield _('Image Inverted.')
-                        except OSError:
-                            yield _('Image type cannot be converted. %s') % img.mode
-                return
-            elif args[0] == 'flip':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = ImageOps.flip(img)
-                        element.altered()
-                        yield _('Image Flipped.')
-                return
-            elif args[0] == 'mirror':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = ImageOps.mirror(img)
-                        element.altered()
-                        yield _('Image Mirrored.')
-                return
-            elif args[0] == 'ccw':
-                from PIL import Image
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.transpose(Image.ROTATE_90)
-                        element.image_height, element.image_width = element.image_width, element.image_height
-                        element.altered()
-                        yield _('Rotated image counterclockwise.')
-                return
-            elif args[0] == 'cw':
-                from PIL import Image
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = img.transpose(Image.ROTATE_270)
-                        element.image_height, element.image_width = element.image_width, element.image_height
-                        element.altered()
-                        yield _('Rotated image clockwise.')
-                return
-            elif args[0] == 'autocontrast':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        try:
-                            cutoff = int(args[1])
-                            img = element.image
-                            if img.mode == 'RGBA':
-                                img = img.convert('RGB')
-                            element.image = ImageOps.autocontrast(img, cutoff=cutoff)
-                            element.altered()
-                            yield _('Image Auto-Contrasted.')
-                        except (IndexError, ValueError):
-                            yield _('image autocontrast <cutoff-percent>')
-                return
-            elif args[0] == 'grayscale' or args[0] == 'greyscale':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = ImageOps.grayscale(img)
-                        element.altered()
-                        yield _('Image Grayscale.')
-                return
-            elif args[0] == 'equalize':
-                from PIL import ImageOps
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        img = element.image
-                        element.image = ImageOps.equalize(img)
-                        element.altered()
-                        yield _('Image Equalized.')
-                return
-            elif args[0] == 'save':
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        try:
-                            img = element.image
-                            img.save(args[1])
-                            yield _('Saved: %s') % (args[1])
-                        except IndexError:
-                            yield _('No file given.')
-                        except OSError:
-                            yield _('File could not be written / created.')
-                        except ValueError:
-                            yield _('Could not determine expected format.')
-
-                return
-            elif args[0] == 'flatrotary':
-                for element in elements.elems(emphasized=True):
-                    if isinstance(element, SVGImage):
-                        points = len(args) - 1
-                        im = element.image
-                        w, h = im.size
-                        from PIL import Image
-
-                        def t(i):
-                            return int(i * w / (points - 1))
-
-                        def x(i):
-                            return int(w * float(args[i + 1]))
-
-                        boxes = list((t(i), 0, t(i + 1), h) for i in range(points - 1))
-                        quads = list((x(i), 0, x(i), h, x(i + 1), h, x(i + 1), 0) for i in range(points - 1))
-                        mesh = list(zip(boxes, quads))
-                        element.image = im.transform(im.size, Image.MESH,
-                                                     mesh,
-                                                     Image.BILINEAR)
-                        element.altered()
-            elif args[0] == 'halftone':
-                #  https://stackoverflow.com/questions/10572274/halftone-images-in-python/10575940#10575940
-                pass
-            else:
-                yield _('Image command unrecognized.')
-                return
-
-        kernel.register('command/image', image)
-
-        def halftone(command, *args):
-            '''Returns list of half-tone images for cmyk image. sample (pixels),
-               determines the sample box size from the original image. The maximum
-               output dot diameter is given by sample * scale (which is also the number
-               of possible dot sizes). So sample=1 will presevere the original image
-               resolution, but scale must be >1 to allow variation in dot size.'''
-            from PIL import Image, ImageDraw, ImageStat
-            oversample = 2
-            sample = 10
-            scale = 1
-            angle = 22
-            for element in elements.elems(emphasized=True):
-                if not isinstance(element, SVGImage):
-                    continue
-                image = element.image
-                im = image
-                image = image.convert('L')
-                image = image.rotate(angle, expand=1)
-                size = image.size[0] * scale, image.size[1] * scale
-                half_tone = Image.new('L', size)
-                draw = ImageDraw.Draw(half_tone)
-                for x in range(0, image.size[0], sample):
-                    for y in range(0, image.size[1], sample):
-                        box = image.crop(
-                            (x - oversample, y - oversample, x + sample + oversample, y + sample + oversample))
-                        stat = ImageStat.Stat(box)
-                        diameter = (stat.mean[0] / 255) ** 0.5
-                        edge = 0.5 * (1 - diameter)
-                        x_pos, y_pos = (x + edge) * scale, (y + edge) * scale
-                        box_edge = sample * diameter * scale
-                        draw.ellipse((x_pos, y_pos, x_pos + box_edge, y_pos + box_edge), fill=255)
-                half_tone = half_tone.rotate(-angle, expand=1)
-                width_half, height_half = half_tone.size
-                xx = (width_half - im.size[0] * scale) / 2
-                yy = (height_half - im.size[1] * scale) / 2
-                half_tone = half_tone.crop((xx, yy, xx + im.size[0] * scale, yy + im.size[1] * scale))
-            element.image = half_tone
-            element.altered()
-
-        kernel.register('command/halftone', halftone)
-        kernel.register('command-help/halftone', 'image halftone <diameter> <scale> <angle> ')
-
-        def trace_hull(command, *args):
-            spooler = context.active.spooler
-            pts = []
-            for obj in elements.elems(emphasized=True):
-                if isinstance(obj, Path):
-                    epath = abs(obj)
-                    pts += [q for q in epath.as_points()]
-                elif isinstance(obj, SVGImage):
-                    bounds = obj.bbox()
-                    pts += [(bounds[0], bounds[1]),
-                            (bounds[0], bounds[3]),
-                            (bounds[2], bounds[1]),
-                            (bounds[2], bounds[3])]
-            hull = [p for p in Point.convex_hull(pts)]
-            if len(hull) == 0:
-                yield _('No elements bounds to trace.')
-                return
-            hull.append(hull[0])  # loop
-
-            def trace_hull():
-                yield COMMAND_WAIT_FINISH
-                yield COMMAND_MODE_RAPID
-                for p in hull:
-                    yield COMMAND_MOVE, p[0], p[1]
-
-            spooler.job(trace_hull)
-            return
-
-        kernel.register('command/trace_hull', trace_hull)
-
-        def trace_quick(command, *args):
-            spooler = context.active.spooler
-            bbox = elements.bounds()
-            if bbox is None:
-                yield _('No elements bounds to trace.')
-                return
-
-            def trace_quick():
-                yield COMMAND_MODE_RAPID
-                yield COMMAND_MOVE, bbox[0], bbox[1]
-                yield COMMAND_MOVE, bbox[2], bbox[1]
-                yield COMMAND_MOVE, bbox[2], bbox[3]
-                yield COMMAND_MOVE, bbox[0], bbox[3]
-                yield COMMAND_MOVE, bbox[0], bbox[1]
-
-            spooler.job(trace_quick)
-            return
-
-        kernel.register('command/trace_quick', trace_quick)
-
-    def add_element(self, element):
-        context_root = self.context.get_context('/')
-        element.stroke = Color('black')
-        context_root.elements.add_elem(element)
-        context_root.elements.set_selected([element])
-
-    def register(self, obj):
-        obj.cache = None
-        obj.icon = None
-        obj.bounds = None
-        obj.last_transform = None
-        obj.selected = False
-        obj.emphasized = False
-        obj.highlighted = False
-
-        def select():
-            obj.selected = True
-            self.context.signal('selected', obj)
-
-        def unselect():
-            obj.selected = False
-            self.context.signal('selected', obj)
-
-        def highlight():
-            obj.highlighted = True
-            self.context.signal('highlighted', obj)
-
-        def unhighlight():
-            obj.highlighted = False
-            self.context.signal('highlighted', obj)
-
-        def emphasize():
-            obj.emphasized = True
-            self.context.signal('emphasized', obj)
-            self.validate_bounds()
-
-        def unemphasize():
-            obj.emphasized = False
-            self.context.signal('emphasized', obj)
-            self.validate_bounds()
-
-        def modified():
-            """
-            The matrix transformation was changed.
-            """
-            obj.bounds = None
-            self._bounds = None
-            self.validate_bounds()
-            self.context.signal('modified', obj)
-
-        def altered():
-            """
-            The data structure was changed.
-            """
-            try:
-                obj.cache.UnGetNativePath(obj.cache.NativePath)
-            except AttributeError:
-                pass
-            del obj.cache
-            obj.cache = None
-            del obj.icon
-            obj.icon = None
-            obj.bounds = None
-            self._bounds = None
-            self.validate_bounds()
-            self.context.signal('altered', obj)
-
-        obj.select = select
-        obj.unselect = unselect
-        obj.highlight = highlight
-        obj.unhighlight = unhighlight
-        obj.emphasize = emphasize
-        obj.unemphasize = unemphasize
-        obj.modified = modified
-        obj.altered = altered
-
-    def unregister(self, e):
-        try:
-            e.cache.UngetNativePath(e.cache.NativePath)
-        except AttributeError:
-            pass
-        try:
-            del e.cache
-        except AttributeError:
-            pass
-        try:
-            del e.icon
-        except AttributeError:
-            pass
-        try:
-            e.unselect()
-            e.unemphasize()
-            e.unhighlight()
-            e.modified()
-        except AttributeError:
-            pass
-
-    def load_default(self):
-        self.clear_operations()
-        self.add_op(LaserOperation(operation="Image", color="black",
-                                   speed=140.0,
-                                   power=1000.0,
-                                   raster_step=3))
-        self.add_op(LaserOperation(operation="Raster", color="black", speed=140.0))
-        self.add_op(LaserOperation(operation="Engrave", color="blue", speed=35.0))
-        self.add_op(LaserOperation(operation="Cut", color="red", speed=10.0))
-        self.classify(self.elems())
-
-    def load_default2(self):
-        self.clear_operations()
-        self.add_op(LaserOperation(operation="Image", color="black",
-                                   speed=140.0,
-                                   power=1000.0,
-                                   raster_step=3))
-        self.add_op(LaserOperation(operation="Raster", color="black", speed=140.0))
-        self.add_op(LaserOperation(operation="Engrave", color="green", speed=35.0))
-        self.add_op(LaserOperation(operation="Engrave", color="blue", speed=35.0))
-        self.add_op(LaserOperation(operation="Engrave", color="magenta", speed=35.0))
-        self.add_op(LaserOperation(operation="Engrave", color="cyan", speed=35.0))
-        self.add_op(LaserOperation(operation="Engrave", color="yellow", speed=35.0))
-        self.add_op(LaserOperation(operation="Cut", color="red", speed=10.0))
-        self.classify(self.elems())
-
-    def finalize(self, channel=None):
-        context = self.context
-        settings = context.derive('operations')
-        settings.clear_persistent()
-        for i, op in enumerate(self.ops()):
-            op_set = settings.derive(str(i))
-            for key in dir(op):
-                if key.startswith('_'):
-                    continue
-                value = getattr(op, key)
-                if value is None:
-                    continue
-                if isinstance(value, Color):
-                    value = value.value
-                op_set.write_persistent(key, value)
-
-    def boot(self, channel=None):
-        settings = self.context.derive('operations')
-        subitems = list(settings.derivable())
-        ops = [None] * len(subitems)
-        for i, v in enumerate(subitems):
-            op_set = settings.derive(v)
-            op = LaserOperation()
-            op_set.load_persistent_object(op)
-            try:
-                ops[i] = op
-            except (ValueError, IndexError):
-                ops.append(op)
-        if not len(ops):
-            self.load_default()
-            return
-        self.add_ops([o for o in ops if o is not None])
-        self.context.signal('rebuild_tree')
-
-    def items(self, **kwargs):
-        def combined(*args):
-            for listv in args:
-                for itemv in listv:
-                    yield itemv
-
-        for j in combined(self.ops(**kwargs), self.elems(**kwargs)):
-            yield j
-
-    def _filtered_list(self, item_list, **kwargs):
-        """
-        Filters a list of items with selected, emphasized, and highlighted.
-        False values means find where that parameter is false.
-        True values means find where that parameter is true.
-        If the filter does not exist then it isn't used to filter that data.
-
-        Items which are set to None are skipped.
-
-        :param item_list:
-        :param kwargs:
-        :return:
-        """
-        s = 'selected' in kwargs
-        if s:
-            s = kwargs['selected']
-        else:
-            s = None
-        e = 'emphasized' in kwargs
-        if e:
-            e = kwargs['emphasized']
-        else:
-            e = None
-        h = 'highlighted' in kwargs
-        if h:
-            h = kwargs['highlighted']
-        else:
-            h = None
-        for obj in item_list:
-            if obj is None:
-                continue
-            if s is not None and s != obj.selected:
-                continue
-            if e is not None and e != obj.emphasized:
-                continue
-            if h is not None and s != obj.highlighted:
-                continue
-            yield obj
-
-    def ops(self, **kwargs):
-        for item in self._filtered_list(self._operations, **kwargs):
-            yield item
-
-    def elems(self, **kwargs):
-        for item in self._filtered_list(self._elements, **kwargs):
-            yield item
-
-    def plan(self, **kwargs):
-        for item in self._plan:
-            yield item
-
-    def add_path(self, path, settings):
-        for seg in path:
-            if isinstance(seg, Move):
-                pass  # Move operations are ignored.
-            elif isinstance(seg, Close):
-                self._plan.append(LineCut(seg.start[0], seg.start[1], seg.end[0], seg.end[1], settings=settings))
-            elif isinstance(seg, Line):
-                self._plan.append(LineCut(seg.start[0], seg.start[1], seg.end[0], seg.end[1], settings=settings))
-            elif isinstance(seg, QuadraticBezier):
-                self._plan.append(QuadCut(seg.start[0], seg.start[1], seg.control[0], seg.control[1],
-                                          seg.end[0], seg.end[1], settings=settings))
-            elif isinstance(seg, CubicBezier):
-                self._plan.append(CubicCut(seg.start[0], seg.start[1], seg.control1[0], seg.control1[1],
-                                           seg.control2[0], seg.control2[1], seg.end[0], seg.end[1], settings=settings))
-            elif isinstance(seg, Arc):
-                arc = ArcCut(seg, settings=settings)
-                self._plan.append(arc)
-                # arc.arc_to(seg)
-
-    def add_raster(self, image, settings):
-        self._plan.append(image, settings)
-
-    def add_plan(self, adding_plan):
-        self._plan.extend(adding_plan)
-
-    def first_element(self, **kwargs):
-        for e in self.elems(**kwargs):
-            return e
-        return None
-
-    def has_emphasis(self):
-        return self.first_element(emphasized=True) is not None
-
-    def count_elems(self, **kwargs):
-        return len(list(self.elems(**kwargs)))
-
-    def count_op(self, **kwargs):
-        return len(list(self.ops(**kwargs)))
-
-    def get_op(self, index, **kwargs):
-        for i, op in enumerate(self.ops(**kwargs)):
-            if i == index:
-                return op
-        raise IndexError
-
-    def get_elem(self, index, **kwargs):
-        for i, elem in enumerate(self.elems(**kwargs)):
-            if i == index:
-                return elem
-        raise IndexError
-
-    def add_op(self, op):
-        self._operations.append(op)
-        self.register(op)
-        self.context.signal('operation_added', op)
-
-    def add_ops(self, adding_ops):
-        self._operations.extend(adding_ops)
-        for op in adding_ops:
-            self.register(op)
-        self.context.signal('operation_added', adding_ops)
-
-    def add_elem(self, element):
-        self._elements.append(element)
-        self.register(element)
-        self.context.signal('element_added', element)
-
-    def add_elems(self, adding_elements):
-        self._elements.extend(adding_elements)
-        for element in adding_elements:
-            self.register(element)
-        self.context.signal('element_added', adding_elements)
-
-    def files(self):
-        return self._filenodes
-
-    def clear_operations(self):
-        for op in self._operations:
-            self.unregister(op)
-            self.context.signal('operation_removed', op)
-        self._operations.clear()
-
-    def clear_elements(self):
-        for e in self._elements:
-            self.unregister(e)
-            self.context.signal('element_removed', e)
-        self._elements.clear()
-
-    def clear_plan(self):
-        self._plan.clear()
-
-    def clear_files(self):
-        self._filenodes.clear()
-
-    def clear_elements_and_operations(self):
-        self.clear_elements()
-        self.clear_operations()
-
-    def clear_all(self):
-        self.clear_plan()
-        self.clear_elements()
-        self.clear_operations()
-        self.clear_files()
-        self.clear_note()
-        self.validate_bounds()
-
-    def clear_note(self):
-        self.note = None
-
-    def remove_files(self, file_list):
-        for f in file_list:
-            del self._filenodes[f]
-
-    def remove_elements(self, elements_list):
-        for elem in elements_list:
-            for i, e in enumerate(self._elements):
-                if elem is e:
-                    self.unregister(elem)
-                    self.context.signal('element_removed', elem)
-                    self._elements[i] = None
-        self.remove_elements_from_operations(elements_list)
-        self.validate_bounds()
-
-    def remove_operations(self, operations_list):
-        for op in operations_list:
-            for i, o in enumerate(self._operations):
-                if o is op:
-                    self.unregister(op)
-                    self.context.signal('operation_removed', op)
-                    self._operations[i] = None
-        self.purge_unset()
-
-    def remove_elements_from_operations(self, elements_list):
-        for i, op in enumerate(self._operations):
-            if op is None:
-                continue
-            elems = [e for e in op if e not in elements_list]
-            op.clear()
-            op.extend(elems)
-        self.purge_unset()
-
-    def purge_unset(self):
-        if None in self._operations:
-            ops = [op for op in self._operations if op is not None]
-            self._operations.clear()
-            self._operations.extend(ops)
-        if None in self._elements:
-            elems = [elem for elem in self._elements if elem is not None]
-            self._elements.clear()
-            self._elements.extend(elems)
-
-    def bounds(self):
-        return self._bounds
-
-    def validate_bounds(self):
-        boundary_points = []
-        for e in self._elements:
-            if e.last_transform is None or e.last_transform != e.transform or e.bounds is None:
-                try:
-                    e.bounds = e.bbox(False)
-                except AttributeError:
-                    # Type does not have bbox.
-                    continue
-                e.last_transform = copy(e.transform)
-            if e.bounds is None:
-                continue
-            if not e.emphasized:
-                continue
-            box = e.bounds
-            top_left = e.transform.point_in_matrix_space([box[0], box[1]])
-            top_right = e.transform.point_in_matrix_space([box[2], box[1]])
-            bottom_left = e.transform.point_in_matrix_space([box[0], box[3]])
-            bottom_right = e.transform.point_in_matrix_space([box[2], box[3]])
-            boundary_points.append(top_left)
-            boundary_points.append(top_right)
-            boundary_points.append(bottom_left)
-            boundary_points.append(bottom_right)
-
-        if len(boundary_points) == 0:
-            new_bounds = None
-        else:
-            xmin = min([e[0] for e in boundary_points])
-            ymin = min([e[1] for e in boundary_points])
-            xmax = max([e[0] for e in boundary_points])
-            ymax = max([e[1] for e in boundary_points])
-            new_bounds = [xmin, ymin, xmax, ymax]
-        if self._bounds != new_bounds:
-            self._bounds = new_bounds
-            self.context.signal('selected_bounds', self._bounds)
-
-    def is_in_set(self, v, selected, flat=True):
-        for q in selected:
-            if flat and isinstance(q, (list, tuple)) and self.is_in_set(v, q, flat):
-                return True
-            if q is v:
-                return True
-        return False
-
-    def set_selected(self, selected):
-        """
-        Sets selected and other properties of a given element.
-
-        All selected elements are also semi-selected.
-
-        If elements itself is selected, all subelements are semiselected.
-
-        If any operation is selected, all sub-operations are highlighted.
-
-        """
-        if selected is None:
-            selected = []
-        for s in self._elements:
-            should_select = self.is_in_set(s, selected, False)
-            should_emphasize = self.is_in_set(s, selected)
-            if s.emphasized:
-                if not should_emphasize:
-                    s.unemphasize()
-            else:
-                if should_emphasize:
-                    s.emphasize()
-            if s.selected:
-                if not should_select:
-                    s.unselect()
-            else:
-                if should_select:
-                    s.select()
-        for s in self._operations:
-            should_select = self.is_in_set(s, selected, False)
-            should_emphasize = self.is_in_set(s, selected)
-            if s.emphasized:
-                if not should_emphasize:
-                    s.unemphasize()
-            else:
-                if should_emphasize:
-                    s.emphasize()
-            if s.selected:
-                if not should_select:
-                    s.unselect()
-            else:
-                if should_select:
-                    s.select()
-
-    def center(self):
-        bounds = self._bounds
-        return (bounds[2] + bounds[0]) / 2.0, (bounds[3] + bounds[1]) / 2.0
-
-    def ensure_positive_bounds(self):
-        b = self._bounds
-        self._bounds = [min(b[0], b[2]), min(b[1], b[3]), max(b[0], b[2]), max(b[1], b[3])]
-        self.context.signal('selected_bounds', self._bounds)
-
-    def update_bounds(self, b):
-        self._bounds = [b[0], b[1], b[0], b[1]]
-        self.context.signal('selected_bounds', self._bounds)
-
-    @staticmethod
-    def bounding_box(elements):
-        if isinstance(elements, SVGElement):
-            elements = [elements]
-        elif isinstance(elements, list):
-            try:
-                elements = [e.object for e in elements if isinstance(e.object, SVGElement)]
-            except AttributeError:
-                pass
-        boundary_points = []
-        for e in elements:
-            box = e.bbox(False)
-            if box is None:
-                continue
-            top_left = e.transform.point_in_matrix_space([box[0], box[1]])
-            top_right = e.transform.point_in_matrix_space([box[2], box[1]])
-            bottom_left = e.transform.point_in_matrix_space([box[0], box[3]])
-            bottom_right = e.transform.point_in_matrix_space([box[2], box[3]])
-            boundary_points.append(top_left)
-            boundary_points.append(top_right)
-            boundary_points.append(bottom_left)
-            boundary_points.append(bottom_right)
-        if len(boundary_points) == 0:
-            return None
-        xmin = min([e[0] for e in boundary_points])
-        ymin = min([e[1] for e in boundary_points])
-        xmax = max([e[0] for e in boundary_points])
-        ymax = max([e[1] for e in boundary_points])
-        return xmin, ymin, xmax, ymax
-
-    def move_selected(self, dx, dy):
-        for obj in self.elems(emphasized=True):
-            obj.transform.post_translate(dx, dy)
-            obj.modified()
-
-    def set_selected_by_position(self, position):
-        def contains(box, x, y=None):
-            if y is None:
-                y = x[1]
-                x = x[0]
-            return box[0] <= x <= box[2] and box[1] <= y <= box[3]
-
-        if self.has_emphasis():
-            if self._bounds is not None and contains(self._bounds, position):
-                return  # Select by position aborted since selection position within current select bounds.
-        for e in reversed(list(self.elems())):
-            try:
-                bounds = e.bbox()
-            except AttributeError:
-                continue  # No bounds.
-            if bounds is None:
-                continue
-            if contains(bounds, position):
-                self.set_selected([e])
-                return
-        self.set_selected(None)
-
-    def classify(self, elements):
-        """
-        Classify does the initial placement of elements as operations.
-        "Image" is the default for images.
-        If element strokes are red they get classed as cut operations
-        If they are otherwise they get classed as engrave.
-        """
-        if elements is None:
-            return
-        for element in elements:
-            found_color = False
-            try:
-                stroke = element.stroke
-            except AttributeError:
-                stroke = None  # Element has no stroke.
-            try:
-                fill = element.fill
-            except AttributeError:
-                fill = None  # Element has no stroke.
-            for op in self.ops():
-                if op.operation in ("Engrave", "Cut", "Raster") and op.color == stroke:
-                    op.append(element)
-                    found_color = True
-                elif op.operation == 'Image' and isinstance(element, SVGImage):
-                    op.append(element)
-                    found_color = True
-                elif op.operation == 'Raster' and \
-                        not isinstance(element, SVGImage) and \
-                        fill is not None and \
-                        fill.value is not None:
-                    op.append(element)
-                    found_color = True
-            if not found_color:
-                if stroke is not None and stroke.value is not None:
-                    op = LaserOperation(operation="Engrave", color=stroke, speed=35.0)
-                    op.append(element)
-                    self.add_op(op)
-
-    def load(self, pathname, **kwargs):
-        kernel = self.context._kernel
-        for loader_name in kernel.match('load'):
-            loader = kernel.registered[loader_name]
-            for description, extensions, mimetype in loader.load_types():
-                if pathname.lower().endswith(extensions):
-                    results = loader.load(self.context, pathname, **kwargs)
-                    if results is None:
-                        continue
-                    elements, ops, note, pathname, basename = results
-                    self._filenodes[pathname] = elements
-                    self.add_elems(elements)
-                    if ops is not None:
-                        self.clear_operations()
-                        self.add_ops(ops)
-                    if note is not None:
-                        self.clear_note()
-                        self.note = note
-                    return elements, pathname, basename
-        return None
-
-    def load_types(self, all=True):
-        kernel = self.context._kernel
-        filetypes = []
-        if all:
-            filetypes.append('All valid types')
-            exts = []
-            for loader_name in kernel.match('load'):
-                loader = kernel.registered[loader_name]
-                for description, extensions, mimetype in loader.load_types():
-                    for ext in extensions:
-                        exts.append('*.%s' % ext)
-            filetypes.append(';'.join(exts))
-        for loader_name in kernel.match('load'):
-            loader = kernel.registered[loader_name]
-            for description, extensions, mimetype in loader.load_types():
-                exts = []
-                for ext in extensions:
-                    exts.append('*.%s' % ext)
-                filetypes.append("%s (%s)" % (description, extensions[0]))
-                filetypes.append(';'.join(exts))
-        return "|".join(filetypes)
-
-    def save(self, pathname):
-        kernel = self.context._kernel
-        for save_name in kernel.match('save'):
-            saver = kernel.registered[save_name]
-            for description, extension, mimetype in saver.save_types():
-                if pathname.lower().endswith(extension):
-                    saver.save(self.context, pathname, 'default')
-                    return True
-        return False
-
-    def save_types(self):
-        kernel = self.context._kernel
-        filetypes = []
-        for save_name in kernel.match('save'):
-            saver = kernel.registered[save_name]
-            for description, extension, mimetype in saver.save_types():
-                filetypes.append("%s (%s)" % (description, extension))
-                filetypes.append("*.%s" % (extension))
-        return "|".join(filetypes)
-
-
 class BindAlias(Modifier):
     """
     Functionality to add BindAlias commands.
@@ -3018,6 +882,7 @@ class Context:
             value = getattr(self, attr)
             if value is None:
                 continue
+            from svgelements import Color
             if isinstance(value, (int, bool, str, float, Color)):
                 self._kernel.write_persistent(self.abs_path(attr), value)
 
@@ -3209,6 +1074,7 @@ class Context:
                 continue
             obj_value = getattr(obj, attr)
 
+            from svgelements import Color
             if not isinstance(obj_value, (int, float, str, bool, Color)):
                 continue
             load_value = self._kernel.read_persistent(type(obj_value), self.abs_path(attr))
@@ -3401,7 +1267,6 @@ class Kernel:
 
     @staticmethod
     def sub_register(device):
-        device.register('modifier/Elemental', Elemental)
         device.register('modifier/Spooler', Spooler)
         device.register('modifier/BindAlias', BindAlias)
 
@@ -3653,6 +1518,7 @@ class Kernel:
 
         def run():
             self.threads[thread_name] = thread
+            # TODO: Restore this.
             # try:
             func()
             # except:
